+++ vnext/metro.config.js
@@ -41,9 +41,7 @@ module.exports = {
   ],
 
   resolver: {
-    // We need a custom resolveRequest right now since our integration tests use a "windesktop" platform thats specific to integration tests.
     resolveRequest: reactNativePlatformResolver({
-      windesktop: 'react-native-windows',
       windows: 'react-native-windows',
     }),
     extraNodeModules: {