+++ packages/microsoft-reactnative-sampleapps/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - microsoft-reactnative-sampleapps
 
-This log was last generated on Fri, 15 Jan 2021 05:07:01 GMT and should not be manually modified.
+This log was last generated on Sun, 17 Jan 2021 05:08:07 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.54
+
+Sun, 17 Jan 2021 05:08:07 GMT
+
+### Changes
+
+- Bump react-native-windows to v0.0.0-canary.237 (asklar@microsoft.com)
+
+## 0.0.54
+
+Sat, 16 Jan 2021 05:05:46 GMT
+
+### Changes
+
+- Bump react-native-windows to v0.0.0-canary.236 (asklar@winse.microsoft.com)
 
 ## 0.0.54